test(e2e-testing): Add eyes snapshots, reduce tests (#21228)
# Overview

Deleted one of the playwright tests. We might have to refactor the plate
reader test at some point.

[Trying to follow step one of the ticket and break it up.
](https://opentrons.atlassian.net/browse/RQA-5262)

Should provide a bit more onboarding flow coverage for when we fork
OT-2.
