--- e2e-testing/tests/pd/test_pd_create_new_flex.py
@@ -1,99 +1,0 @@
-"""Test Flex robot protocol creation workflow.
-
-Ports tests from protocol-designer/cypress/e2e/createNewFlex.cy.ts.
-
-This test verifies the complete Flex onboarding workflow including:
-- Robot selection (OT-2 <-> Flex switching)
-- Pipette configuration
-- Gripper, thermocycler, and waste chute selection
-- Module configuration (thermocycler, heater-shaker, mag block, temp module)
-- Deck configuration
-- Adding labware to deck
-- Liquid definition and assignment
-"""
-
-import re
-
-import pytest
-from playwright.sync_api import Page, expect
-
-from automation.pd_pages import LandingPage
-
-
-@pytest.mark.pdE2E
-@pytest.mark.slow
-def test_flex_onboarding_workflow(page: Page, pd_base_url: str) -> None:
-    """Test complete Flex onboarding workflow.
-
-    This test replicates the Cypress createNewFlex.cy.ts test:
-    1. Start protocol creation
-    2. Step 1: Verify Flex is selected by default, switch to OT-2, then back to Flex
-    3. Step 2: Configure 50µL single-channel pipette
-    4. Step 3: Select gripper, no thermocycler, no waste chute
-    5. Step 4: Add modules to deck (thermocycler, heater-shaker, mag block, temp module)
-    6. Continue to protocol editor
-    """
-    # Start on home page
-    landing = LandingPage(page)
-    landing.wait_for_page_load()
-    landing.confirm_welcome_modal()
-
-    # Click "Create a protocol" to start onboarding
-    landing.click_create_protocol()
-
-    # Step 1: Complete all basic questions (robot, pipettes, gripper, modules)
-    expect(page.get_by_text("Step 1")).to_be_visible()
-
-    # Verify Flex is available (it's selected by default)
-    expect(page.locator("label", has_text="Opentrons Flex")).to_be_visible()
-
-    # Part 1: Add pipette - Click "Add a pipette" to open selector
-    page.get_by_text("Add a pipette").click()
-
-    # Select 1-Channel pipette type
-    page.get_by_text("1-Channel").click()
-
-    # Select 50 µL volume
-    page.get_by_text("50 µL").click()
-
-    # Select Tip Rack 50 µL (not Filter Tip Rack) - use exact match
-    page.locator("label").filter(has=page.get_by_text(re.compile(r"^Tip Rack 50 µL$"))).first.click()
-
-    # Save pipette configuration
-    page.get_by_role("button", name="Save").click()
-
-    # Part 2: Gripper question - "Do you want to move labware automatically with the gripper?"
-    expect(page.get_by_text("Do you want to move labware automatically with the gripper?")).to_be_visible()
-    page.get_by_text("Yes", exact=True).click()
-
-    # Part 3: Thermocycler question - "Are you using a Thermocycler in your protocol?"
-    expect(page.get_by_text("Are you using a Thermocycler in your protocol?")).to_be_visible()
-    page.get_by_test_id("BasicsButtons_thermocycler_no").get_by_text("No").click()
-
-    # Part 4: Waste chute question - "Are you using a waste chute in your protocol?"
-    expect(page.get_by_text("Are you using a waste chute in your protocol?")).to_be_visible()
-    page.get_by_test_id("BasicsButtons_wasteChute_no").get_by_text("No").click()
-
-    # Now Confirm button should be enabled - click to proceed to Step 2
-    confirm_button = page.get_by_role("button", name="Confirm")
-    expect(confirm_button).to_be_enabled()
-    confirm_button.click()
-
-    # Step 2: Configure deck hardware
-    expect(page.get_by_text("Step 2")).to_be_visible(timeout=10000)
-    expect(page.get_by_text("Configure your deck hardware")).to_be_visible()
-
-    # Test successfully reached Step 2 of Flex onboarding
-    # This verifies:
-    # 1. Robot selection (Flex)
-    # 2. Pipette configuration (1-Channel 50µL with Tip Rack 50µL)
-    # 3. Gripper selection (Yes)
-    # 4. Thermocycler selection (No)
-    # 5. Waste chute selection (No)
-    # 6. Progression to deck hardware configuration
-
-    # Note: Full module addition workflow would require:
-    # - Clicking deck slots (e.g., data-testid="D1")
-    # - Selecting module category (e.g., "Modules")
-    # - Choosing specific module (e.g., "Heater-Shaker Module GEN1")
-    # This is left for future enhancement as it requires more complex page object methods
